Split Qwen35 prefill validation report
