+++ crates/runner/src/main.rs
@@ -28,6 +28,7 @@ mod qwen35_engine_setup;
 mod qwen35_kv_trace;
 mod qwen35_oracle_prefill_trace;
 mod qwen35_prefill;
+mod qwen35_prefill_validation_report;
 mod qwen35_runtime;
 mod qwen35_startup;
 mod qwen35_trace;
